+++ testng/src/main/java/cucumber/api/testng/TestNGCucumberRunner.java
@@ -1,102 +1,55 @@
 package cucumber.api.testng;
 
-import cucumber.api.event.TestRunFinished;
-import cucumber.api.event.TestRunStarted;
-import cucumber.runner.*;
-import cucumber.runtime.*;
-import cucumber.runtime.filter.Filters;
-import cucumber.runtime.formatter.PluginFactory;
-import cucumber.runtime.formatter.Plugins;
-import cucumber.runtime.io.MultiLoader;
-import cucumber.runtime.io.ResourceLoader;
-import cucumber.runtime.io.ResourceLoaderClassFinder;
-import cucumber.runtime.model.CucumberFeature;
-import cucumber.runtime.model.FeatureLoader;
 import gherkin.events.PickleEvent;
-
-import java.util.ArrayList;
-import java.util.List;
+import io.cucumber.testng.CucumberFeatureWrapper;
+import io.cucumber.testng.PickleEventWrapper;
+import org.apiguardian.api.API;
 
 /**
  * Glue code for running Cucumber via TestNG.
+ *
+ * @deprecated use {@code io.cucumber.testng.TestNGCucumberRunner} instead
  */
+@API(status = API.Status.MAINTAINED)
+@Deprecated
 public class TestNGCucumberRunner {
-    private final EventBus bus;
-    private final Filters filters;
-    private final FeaturePathFeatureSupplier featureSupplier;
-    private final ThreadLocalRunnerSupplier runnerSupplier;
-    private final RuntimeOptions runtimeOptions;
-    private final Plugins plugins;
 
-    /**
-     * Bootstrap the cucumber runtime
-     *
-     * @param clazz Which has the {@link cucumber.api.CucumberOptions}
-     *              and {@link org.testng.annotations.Test} annotations
-     */
+    private final io.cucumber.testng.TestNGCucumberRunner delegate;
+
     public TestNGCucumberRunner(Class clazz) {
-        ClassLoader classLoader = clazz.getClassLoader();
-        ResourceLoader resourceLoader = new MultiLoader(classLoader);
-
-        RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(clazz);
-        runtimeOptions = runtimeOptionsFactory.create();
-
-        ClassFinder classFinder = new ResourceLoaderClassFinder(resourceLoader, classLoader);
-        BackendModuleBackendSupplier backendSupplier = new BackendModuleBackendSupplier(resourceLoader, classFinder, runtimeOptions);
-        bus = new TimeServiceEventBus(TimeService.SYSTEM);
-        plugins = new Plugins(classLoader, new PluginFactory(), runtimeOptions);
-        FeatureLoader featureLoader = new FeatureLoader(resourceLoader);
-        filters = new Filters(runtimeOptions);
-        this.runnerSupplier = new ThreadLocalRunnerSupplier(runtimeOptions, bus, backendSupplier);
-        featureSupplier = new FeaturePathFeatureSupplier(featureLoader, runtimeOptions);
+        this.delegate = new io.cucumber.testng.TestNGCucumberRunner(clazz);
     }
 
     public void runScenario(PickleEvent pickle) throws Throwable {
-        //Possibly invoked in a multi-threaded context
-        Runner runner = runnerSupplier.get();
-        TestCaseResultListener testCaseResultListener = new TestCaseResultListener(runner.getBus(), runtimeOptions.isStrict());
-        runner.runPickle(pickle);
-        testCaseResultListener.finishExecutionUnit();
-
-        if (!testCaseResultListener.isPassed()) {
-            throw testCaseResultListener.getError();
-        }
+        delegate.runScenario(pickle);
     }
 
     public void finish() {
-        bus.send(new TestRunFinished(bus.getTime(), bus.getTimeMillis()));
+        delegate.finish();
     }
 
-    /**
-     * @return returns the cucumber scenarios as a two dimensional array of {@link PickleEventWrapper}
-     * scenarios combined with their {@link CucumberFeatureWrapper} feature.
-     */
     public Object[][] provideScenarios() {
-        try {
-            List<Object[]> scenarios = new ArrayList<Object[]>();
-            List<CucumberFeature> features = getFeatures();
-            for (CucumberFeature feature : features) {
-                for (PickleEvent pickle : feature.getPickles()) {
-                    if (filters.matchesFilters(pickle)) {
-                        scenarios.add(new Object[]{new PickleEventWrapperImpl(pickle),
-                            new CucumberFeatureWrapperImpl(feature)});
-                    }
+        return addLegacyWrappers(delegate.provideScenarios());
+    }
+
+    private Object[][] addLegacyWrappers(Object[][] objects) {
+        for (Object[] row : objects) {
+            for (int i = 0; i < row.length; i++) {
+                Object element = row[i];
+                if (element == null) {
+                    continue;
+                }
+
+                if (element instanceof PickleEventWrapper) {
+                    PickleEventWrapper wrapper = (PickleEventWrapper) element;
+                    row[i] = new PickleEventWrapperImpl(wrapper);
+                } else if (element instanceof CucumberFeatureWrapper) {
+                    CucumberFeatureWrapper wrapper = (CucumberFeatureWrapper) element;
+                    row[i] = new CucumberFeatureWrapperImpl(wrapper);
                 }
             }
-            return scenarios.toArray(new Object[][]{});
-        } catch (CucumberException e) {
-            return new Object[][]{new Object[]{new CucumberExceptionWrapper(e), null}};
         }
-    }
 
-    List<CucumberFeature> getFeatures() {
-        plugins.setSerialEventBusOnEventListenerPlugins(bus);
-
-        List<CucumberFeature> features = featureSupplier.get();
-        bus.send(new TestRunStarted(bus.getTime(), bus.getTimeMillis()));
-        for (CucumberFeature feature : features) {
-            feature.sendTestSourceRead(bus);
-        }
-        return features;
+        return objects;
     }
 }