+++ testng/src/main/java/io/cucumber/testng/AbstractTestNGCucumberTests.java
@@ -1,4 +1,4 @@
-package cucumber.api.testng;
+package io.cucumber.testng;
 
 import org.apiguardian.api.API;
 import org.testng.annotations.AfterClass;
@@ -8,16 +8,13 @@ import org.testng.annotations.Test;
 
 /**
  * Runs each cucumber scenario found in the features as separated test
- *
- * @deprecated use {@link io.cucumber.testng.AbstractTestNGCucumberTests} instead.
  */
-@Deprecated
-@API(status = API.Status.MAINTAINED)
+@API(status = API.Status.STABLE)
 public abstract class AbstractTestNGCucumberTests {
     private TestNGCucumberRunner testNGCucumberRunner;
 
     @BeforeClass(alwaysRun = true)
-    public void setUpClass() throws Exception {
+    public void setUpClass() {
         testNGCucumberRunner = new TestNGCucumberRunner(this.getClass());
     }
 
@@ -28,7 +25,8 @@ public abstract class AbstractTestNGCucumberTests {
     }
 
     /**
-     * Returns two dimensional array of PickleEventWrapper scenarios with their associated CucumberFeatureWrapper feature.
+     * Returns two dimensional array of PickleEventWrapper scenarios
+     * with their associated CucumberFeatureWrapper feature.
      *
      * @return a two dimensional array of scenarios features.
      */
@@ -41,7 +39,7 @@ public abstract class AbstractTestNGCucumberTests {
     }
 
     @AfterClass(alwaysRun = true)
-    public void tearDownClass() throws Exception {
+    public void tearDownClass() {
         if (testNGCucumberRunner == null) {
             return;
         }