+++ testng/src/main/java/io/cucumber/testng/PickleEventWrapper.java
@@ -1,0 +1,17 @@
+package io.cucumber.testng;
+
+import gherkin.events.PickleEvent;
+import org.apiguardian.api.API;
+
+/**
+ * The only purpose of this interface is to be able to provide a custom
+ * <pre>toString()</pre>, making TestNG reports look more descriptive.
+ *
+ * @see AbstractTestNGCucumberTests#runScenario(PickleEventWrapper, CucumberFeatureWrapper)
+ */
+@API(status = API.Status.STABLE)
+public interface PickleEventWrapper {
+
+    PickleEvent getPickleEvent();
+
+}