+++ testng/src/main/java/io/cucumber/testng/PickleEventWrapperImpl.java
@@ -1,0 +1,21 @@
+package io.cucumber.testng;
+
+import gherkin.events.PickleEvent;
+
+class PickleEventWrapperImpl implements PickleEventWrapper {
+
+    private final PickleEvent pickleEvent;
+
+    PickleEventWrapperImpl(PickleEvent pickleEvent) {
+        this.pickleEvent = pickleEvent;
+    }
+
+    public PickleEvent getPickleEvent() {
+        return pickleEvent;
+    }
+
+    @Override
+    public String toString() {
+        return "\"" + pickleEvent.pickle.getName() + "\"";
+    }
+}