+++ testng/src/test/java/io/cucumber/testng/RunFeatureWithThreeScenariosTest.java
@@ -1,0 +1,10 @@
+package io.cucumber.testng;
+
+
+import cucumber.api.CucumberOptions;
+
+@CucumberOptions(
+    features = "classpath:io/cucumber/testng/three_scenarios.feature"
+)
+public class RunFeatureWithThreeScenariosTest extends AbstractTestNGCucumberTests {
+}